+++ app/(nav)/waitingRoomList/page.tsx
@@ -1,32 +1,16 @@
 'use client';
 
 import TopBar from '@/components/Common/TopBar/TopBar';
-import RoomButton from '@/components/Common/RoomButton/RoomButton';
 import { WaitingRoomListPageWrapper } from '@/app/commonPage.styles';
 import { RoomButtonGrid, ListTitle, SubTitle } from './waitingRoomList.styles';
 import { RefreshIcon } from '@/components/Common/RoomButton/WaitingRoomListIcon';
-import { useEffect } from 'react';
 import useGetWaitingRoomList from '@/hooks/waitingRoomList/useGetWaitingRoomList';
+import DotLoadingSpinner from '@/components/Common/LoadingSpinner/DotLoadingSpinner';
+import RoomButtonList from '@/components/Common/RoomButton/RoomButtonList';
 
 const WaitingRoomList = () => {
-  const { waitingRoomList, fetchWaitingRoomList } = useGetWaitingRoomList();
-  useEffect(() => {
-    const handleWaitingRoomList = async () => {
-      await fetchWaitingRoomList();
-    };
-    handleWaitingRoomList();
-  }, []);
-
-  const roomButtons = waitingRoomList?.map((item, index) => (
-    <RoomButton
-      key={`key${index}`}
-      hostType={'basic'}
-      title={item.roomName}
-      round={item.totalRounds}
-      currMemberCount={item.currentPlayersCount}
-      maxMemberCount={item.maxPlayer}
-    />
-  ));
+  const { waitingRoomList, isLoading, fetchWaitingRoomList } =
+    useGetWaitingRoomList();
 
   return (
     <>
@@ -36,7 +20,13 @@ const WaitingRoomList = () => {
           <ListTitle>참여 가능한 방</ListTitle>
           <RefreshIcon onClick={fetchWaitingRoomList} />
         </SubTitle>
-        <RoomButtonGrid>{roomButtons}</RoomButtonGrid>
+        <RoomButtonGrid>
+          {isLoading ? (
+            <DotLoadingSpinner />
+          ) : (
+            <RoomButtonList waitingRoomList={waitingRoomList} />
+          )}
+        </RoomButtonGrid>
       </WaitingRoomListPageWrapper>
     </>
   );
